Order Suggestions: manual order entry (single + bulk CSV import) + fix a live Demand Adjustment crash
- Add the last deferred piece of Order Suggestions' frontend: a "+ Add
  Manual Order" modal (vendor -> material picker, qty/dates/PO/status/
  notes) and a "Bulk Import (CSV)" modal (paste rows, preview, per-row
  success/failure report), against the already-built
  PurchaseOrderSuggestionHelper.ManualAsync/ManualBulkAsync. This was the
  one remaining "deliberately not ported" item flagged on this tile.
- Continuing the SQL audit: found VendorMasterDataHelper's private
  OverlappingAdjustment record declared AdjustmentId as long against
  log.DemandAdjustment.AdjustmentId's real int column — a guaranteed
  Dapper strict-materialization crash on every single Create/Update
  Demand Adjustment call (FindOverlappingAdjustmentAsync always runs,
  whether or not an overlap is actually found). Fixed to int.

+++ dotnet/NormantonNexus/Models/Dto/VendorMasterDataModels.cs
@@ -49,4 +49,5 @@ public sealed record DemandAdjustmentRow(
 /// </summary>
 public sealed record UpsertDemandAdjustmentRequest(string Material, DateTime? StartDate, DateTime? EndDate, decimal? UsagePercent, string? Reason, decimal? OverrideQty = null);
 
-internal sealed record OverlappingAdjustment(long AdjustmentId, DateTime? StartDate, DateTime? EndDate);
+/// <summary>AdjustmentId is int, matching log.DemandAdjustment.AdjustmentId's real column type — was long, which throws Dapper's strict-materialization error on every Create/Update Demand Adjustment call (FindOverlappingAdjustmentAsync runs on both paths regardless of whether an overlap is actually found).</summary>
+internal sealed record OverlappingAdjustment(int AdjustmentId, DateTime? StartDate, DateTime? EndDate);
